requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.5
django-cleanup==5.1.0
django-crispy-forms==1.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
moviepy==1.0.3
numpy==1.19.5
Pillow==8.1.0
pkg-resources==0.0.0
proglog==0.1.9
pymediainfo==5.0.3
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
tqdm==4.55.1
urllib3==1.26.2
